refactor Rule 110 code to its own file
